+++ spead2/test/test_send_trollius.py
@@ -42,7 +42,7 @@ class TestUdpStream(object):
 
     def test_async_flush(self):
         for i in range(3):
-            trollius.async(self.stream.async_send_heap(self.heap))
+            trollius.ensure_future(self.stream.async_send_heap(self.heap))
         # The above only queues up the async sends on the event loop. The rest of the
         # test needs to be run from inside the event loop
         trollius.get_event_loop().run_until_complete(self._test_async_flush())
@@ -52,7 +52,7 @@ class TestUdpStream(object):
         This is arranged by filling up the queue slots first.
         """
         for i in range(5):
-            trollius.async(self.stream.async_send_heap(self.heap))
+            trollius.ensure_future(self.stream.async_send_heap(self.heap))
         # The above only queues up the async sends on the event loop. The rest of the
         # test needs to be run from inside the event loop
         trollius.get_event_loop().run_until_complete(self._test_async_flush())